# Fix VS Source and Cache paths are parsed improperly