@BoatsMcGee 2024-07-23
This is a polite notice asking you to revert [this commit](https://github.com/vapoursynth/bestsource/commit/cf1fa44dff7c68e1b229ac2aa68f98088d643dbf). Instead, submit the changes collaboratively via Draft Pull Request #60.

PR #60 includes similar code, albeit slightly earlier in execution from that of this commit. Although this commit is not a verbatim copy of my changes, it achieves a veery similar effect. Additionally, PR #60 is a draft to which I explicitly invited collaboration and review as well as stated the following:

> Before completing this PR, please consider the following:
There are temporary logging statements to help demonstrate the issue and how it is resolved should you wish to debug it. **These should be removed or modified to fit the logging practices of the project before merging.**
...
Cache Path is not ensured to be a writable path. **We should discuss where and how this could be handled.** I suggest that if it fails to write to the disk then we should fallback to the default behavior. This should also be reflected in the documentation.

I also provided a summary in your Discord server stating "There is commented code to instead return the path as-is should you not wish to implement a fallback. Other concerns and improvement opportunities have been raised in the PR." This can also be seen along with other context in screenshots below:

However, instead of discussing with me on how to make the Pull Request suitable for your requirements you instead closed it without discussion. You have also halted discussion with me on your private Discord server as shown below:

![PR Closed 0](https://github.com/user-attachments/assets/61d73667-69f5-45b8-a00a-432ba94cb386)
![PR Closed 1](https://github.com/user-attachments/assets/2278d13e-ba73-48c8-91d0-9a4322c26126)
![PR Closed 2](https://github.com/user-attachments/assets/a04e681c-92c4-4239-87e8-f02459f8709a)
![PR Closed 3](https://github.com/user-attachments/assets/2c410300-d003-41b9-8fdb-8752444ae379)
![Uhhh](https://github.com/user-attachments/assets/f250f936-3444-40b8-9cdd-235129480a22)

I will reiterate/rephrase from my last message: I am not insisting that all changes in PR #60 be accepted wholesale. You are free to define and require the necessary modifications to make it befitting of the project.

The Pull Request is a *Draft* Pull Request to facilitate collaboration and requires agreement from both parties before merging. Disagreement with only 1 of many parts of the PR does not justify denying all of it and committing similar changes separately without proper attribution. In other words, if you did not like one of the changes introduced in PR #60, then ask for it to be removed/modified - do not take one of the other changes and submit it as if it were entirely your own.

As this is a polite notice, I will give you benefit of the doubt and assume this commit was made inadvertently without remembering where it originated. Please revert this commit and open PR #60. I will in turn allow you to make the necessary changes.